Move PyramidalService and PyramidalServiceTest to package sc.fiji.ome.zarr.pyramid.fiji
