+++ src/main/java/sc/fiji/ome/zarr/util/BdvUtils.java
@@ -47,6 +47,7 @@ import bdv.util.BdvHandle;
 import bdv.util.BdvOptions;
 import bdv.viewer.SourceAndConverter;
 import sc.fiji.ome.zarr.pyramid.fiji.PyramidalBdv;
+import sc.fiji.ome.zarr.pyramid.fiji.PyramidalService;
 import sc.fiji.ome.zarr.pyramid.metadata.Omero;
 
 public class BdvUtils
