+++ src/test/java/sc/fiji/ome/zarr/plugins/OpenInBDVCommandTest.java
@@ -48,7 +48,7 @@ import org.scijava.command.CommandService;
 
 import sc.fiji.ome.zarr.open.ZarrOpenActions;
 import sc.fiji.ome.zarr.pyramid.Pyramidal;
-import sc.fiji.ome.zarr.util.PyramidalService;
+import sc.fiji.ome.zarr.pyramid.fiji.PyramidalService;
 import sc.fiji.ome.zarr.util.ZarrTestUtils;
 
 class OpenInBDVCommandTest
